skipRequestLayout is no longer needed (and deprecated), some small refactoring

+++ additive_animations/src/main/java/at/wirecube/additiveanimations/helper/propertywrappers/ElevationProperties.java
@@ -27,7 +27,7 @@ public class ElevationProperties {
     public static Property<View, Float> ELEVATION = new FloatProperty<View>("ELEVATION") {
         @Override
         public Float get(View object) {
-            return Float.valueOf(object.getElevation());
+            return object.getElevation();
         }
 
         @Override
